Adding Zip of Customer Relationship Management
In the zip project you just need to add the required jars in the lib as I was not able to push the ZIP file with jars as it was exceeding Github 25mb file size limit. the jars can be found in the above mentioned project itself which I have pushed(web-customer-tracker/src/main/webapp/WEB-INF/lib).